[UI/#97] 학교, 학과, 학번 combine
