+++ app/src/main/java/com/el/yello/presentation/onboarding/fragment/universityinfo/school/SearchDialogSchoolFragment.kt
@@ -1,4 +1,4 @@
-package com.el.yello.presentation.onboarding.fragment.studentid.dialog.department
+package com.el.yello.presentation.onboarding.fragment.universityinfo.school
 
 import android.app.Dialog
 import android.content.Intent
@@ -9,8 +9,10 @@ import android.view.View
 import android.view.WindowManager
 import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.activityViewModels
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.el.yello.R
-import com.el.yello.databinding.FragmentDialogDepartmentBinding
+import com.el.yello.databinding.FragmentDialogSchoolBinding
 import com.el.yello.presentation.onboarding.activity.OnBoardingViewModel
 import com.el.yello.util.context.yelloSnackbar
 import com.example.ui.base.BindingBottomSheetDialog
@@ -19,21 +21,24 @@ import com.example.ui.view.UiState
 import com.example.ui.view.setOnSingleClickListener
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
-import timber.log.Timber
 
-class SearchDialogDepartmentFragment :
-    BindingBottomSheetDialog<FragmentDialogDepartmentBinding>(R.layout.fragment_dialog_department) {
+class SearchDialogSchoolFragment :
+    BindingBottomSheetDialog<FragmentDialogSchoolBinding>(R.layout.fragment_dialog_school) {
+    private var adapter: SchoolAdapter? = null
     private val viewModel by activityViewModels<OnBoardingViewModel>()
-    private var adapter: DepartmentAdapter? = null
+
+    // TODO : viewmodel 이동
+    private var inputText: String = ""
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.vm = viewModel
 
-        initDepartmentAdapter()
-        setupDepartmentData()
+        initView()
+        setupSchoolData()
+        setListWithInfinityScroll()
         recyclerviewScroll()
-        setClicktoDepartmentform()
+        setClickToSchoolForm()
     }
 
     // 바텀 시트 fullScreen
@@ -58,32 +63,50 @@ class SearchDialogDepartmentFragment :
         bottomSheet.layoutParams = layoutParams
     }
 
-    private fun initDepartmentAdapter() {
+    private fun initView() {
         setHideKeyboard()
-        binding.etDepartmentSearch.doAfterTextChanged { input ->
-            viewModel.getGroupList(input.toString())
+        binding.etSchoolSearch.doAfterTextChanged { input ->
+            inputText = input.toString()
+            viewModel.getSchoolList(inputText)
         }
-        adapter = DepartmentAdapter(storeDepartment = ::storeGroup)
-        binding.rvDepartmentList.adapter = adapter
+        adapter = SchoolAdapter(storeSchool = ::storeSchool)
+        binding.rvSchoolList.adapter = adapter
         binding.btnBackDialog.setOnSingleClickListener {
             dismiss()
         }
     }
 
+    private fun setListWithInfinityScroll() {
+        binding.rvSchoolList.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
+                super.onScrolled(recyclerView, dx, dy)
+                if (dy > 0) {
+                    recyclerView.layoutManager?.let { layoutManager ->
+                        if (!binding.rvSchoolList.canScrollVertically(1) &&
+                            layoutManager is LinearLayoutManager &&
+                            layoutManager.findLastVisibleItemPosition() == adapter!!.itemCount - 1
+                        ) {
+                            viewModel.getSchoolList(inputText)
+                        }
+                    }
+                }
+            }
+        })
+    }
+
     private fun setHideKeyboard() {
-        binding.layoutDepartmentDialog.setOnSingleClickListener {
+        binding.layoutSchoolDialog.setOnSingleClickListener {
             requireContext().hideKeyboard(
                 requireView(),
             )
         }
     }
 
-    private fun setupDepartmentData() {
-        viewModel.departmentData.observe(viewLifecycleOwner) { state ->
-            Timber.d("GET GROUP LIST OBSERVE : $state")
+    private fun setupSchoolData() {
+        viewModel.schoolData.observe(viewLifecycleOwner) { state ->
             when (state) {
                 is UiState.Success -> {
-                    adapter?.submitList(state.data.groupList)
+                    adapter?.submitList(state.data.schoolList)
                 }
                 is UiState.Failure -> {
                     yelloSnackbar(binding.root, getString(R.string.msg_error))
@@ -94,38 +117,38 @@ class SearchDialogDepartmentFragment :
         }
     }
 
-    private fun storeGroup(department: String, groupId: Long) {
-        viewModel.setGroupInfo(department, groupId)
-        viewModel.clearDepartmentData()
+    private fun storeSchool(school: String) {
+        viewModel.setSchool(school)
+        viewModel.clearSchoolData()
         dismiss()
     }
     private fun recyclerviewScroll() {
-        binding.rvDepartmentList.setOnTouchListener { view, motionEvent ->
+        binding.rvSchoolList.setOnTouchListener { view, motionEvent ->
             when (motionEvent.action) {
                 MotionEvent.ACTION_MOVE -> {
-                    binding.layoutDepartmentDialog.requestDisallowInterceptTouchEvent(true)
+                    binding.layoutSchoolDialog.requestDisallowInterceptTouchEvent(true)
                 }
             }
             return@setOnTouchListener false
         }
     }
 
-    private fun setClicktoDepartmentform() {
-        binding.tvDepartmentAdd.setOnClickListener {
-            Intent(Intent.ACTION_VIEW, Uri.parse(DEPARTMENT_FORM_URL)).apply {
+    private fun setClickToSchoolForm() {
+        binding.tvSchoolAdd.setOnClickListener {
+            Intent(Intent.ACTION_VIEW, Uri.parse(SCHOOL_FORM_URL)).apply {
                 startActivity(this)
             }
         }
     }
 
     override fun onDestroyView() {
-        adapter = null
         super.onDestroyView()
+        adapter = null
     }
 
     companion object {
         @JvmStatic
-        fun newInstance() = SearchDialogDepartmentFragment()
-        private const val DEPARTMENT_FORM_URL = "https://bit.ly/3pO0ijD"
+        fun newInstance() = SearchDialogSchoolFragment()
+        private const val SCHOOL_FORM_URL = "https://bit.ly/46Yv0Hc"
     }
 }
