--- app/src/main/res/layout/fragment_studentid.xml
@@ -1,113 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="com.el.yello.presentation.onboarding.activity.OnBoardingViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/layout_studentid"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/black">
-
-        <ImageView
-            android:id="@+id/btn_studentid_back_btn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:background="@null"
-            android:padding="12dp"
-            android:src="@drawable/ic_onboarding_btn_back"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/tv_department_where"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginTop="68dp"
-            android:text="@string/onboarding_tv_what_department"
-            android:textAppearance="?textAppearanceHeadline2"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/tv_department_search"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginTop="12dp"
-            android:gravity="center_vertical"
-            android:fontFamily="@font/font_pretendard_medium"
-            android:textSize="16sp"
-            android:background="@{vm._department.isEmpty() ?@drawable/shape_black_fill_grayscales600_line_8_rect : @drawable/shape_grayscales900_fill_grayscales500_line_8_rect}"
-            android:paddingVertical="16dp"
-            android:paddingStart="20dp"
-            android:paddingEnd="12dp"
-            android:textColor="@color/white"
-            app:drawableEndCompat="@drawable/ic_onboarding_search"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tv_department_where" />
-
-        <TextView
-            android:id="@+id/tv_studentid_what"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginTop="76dp"
-            android:text="@string/onboarding_tv_what_studentid"
-            android:textAppearance="?textAppearanceHeadline2"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_department_search" />
-
-        <TextView
-            android:id="@+id/tv_studentid_search"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginTop="12dp"
-            android:background="@{vm._studentId == null ? @drawable/shape_black_fill_grayscales600_line_8_rect : @drawable/shape_grayscales900_fill_grayscales500_line_8_rect}"
-            android:fontFamily="@font/font_pretendard_medium"
-            android:gravity="center_vertical"
-            android:hint="@string/onboarding_tv_hint_select_studentid"
-            android:paddingVertical="16dp"
-            android:paddingStart="20dp"
-            android:paddingEnd="12dp"
-            android:textColor="@color/white"
-            android:textColorHint="@color/grayscales_500"
-            android:textSize="16sp"
-            app:drawableEndCompat="@drawable/ic_onboarding_studentid_down"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tv_studentid_what"
-            tools:background="shape_black_fill_grayscales600_line_8_rect" />
-
-        <android.widget.Button
-            android:id="@+id/btn_stdentid_next"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginBottom="34dp"
-            android:background="@{vm._department.isEmpty() || vm._studentId == null? @drawable/shape_grayscales_800_fill_100_rect : @drawable/shape_yello_main_500_fill_100_rect}"
-            android:clickable="@{!vm._department.isEmpty() &amp; vm._studentId != null}"
-            android:paddingVertical="12dp"
-            android:text="@string/onboarding_btn_next"
-            android:textAppearance="?textAppearanceBodyMedium"
-            android:textColor="@{vm._department.isEmpty()|| vm._studentId == null? @color/grayscales_700 : @color/black}"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
